#1 - removed ApacheHttpClient as obsolete
removed ApacheHttpClient and merged AbstractClient functionality with UrlConnectionClient.
merged tests from AbstractClientUnmarshallingSpockTest to urlConnectionClientSpockTest
