+++ src/test/groovy/com/twingly/search/client/UrlConnectionClientSpockTest.groovy
@@ -9,16 +9,20 @@ import spock.lang.Specification
 
 import javax.xml.bind.JAXBContext
 import javax.xml.bind.JAXBException
+import javax.xml.bind.UnmarshalException
 import javax.xml.bind.Unmarshaller
+import java.nio.file.Paths
 import java.text.SimpleDateFormat
 
 class UrlConnectionClientSpockTest extends Specification {
+    def packagePath = Paths.get("com", "twingly", "search", "client")
     def sdf = new SimpleDateFormat(Constants.DATE_FORMAT)
     def jaxbContext = Mock(JAXBContext)
     def unmarschaller = Mock(Unmarshaller)
     def urlStreamHandler
     def urlConnection = Mock(URLConnection)
     def inputStream = Mock(InputStream)
+    def client = Spy(UrlConnectionClient)
 
     def setup() {
         jaxbContext.createUnmarshaller() >> unmarschaller
@@ -31,10 +35,147 @@ class UrlConnectionClientSpockTest extends Specification {
         urlConnection.getInputStream() >> inputStream
     }
 
+
+    def "should parse valid result result"() {
+        given:
+        def filepath = packagePath.resolve("valid_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        result.numberOfMatchesReturned == 1000
+        result.numberOfMatchesTotal == 60768
+        result.secondsElapsed == 0.022d
+        result.posts.size() == 1000
+    }
+
+    def "should parse valid non blog result result"() {
+        given:
+        def filepath = packagePath.resolve("valid_non_blog_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        result.numberOfMatchesReturned == 2
+        result.numberOfMatchesTotal == 2
+        result.secondsElapsed == 0.022d
+        result.posts.size() == 2
+        result.posts[0].url == "http://www.someurl.com/post"
+        result.posts[0].blogName == "Newspaper Name"
+        result.posts[0].blogUrl == "http://www.someurl.com/"
+        result.posts[0].tags == null
+        result.posts[0].contentType == "newspaper"
+
+        result.posts[1].url == "http://www.someotherurl.com/post"
+        result.posts[1].blogName == "Blog Name"
+        result.posts[1].blogUrl == "http://www.someotherurl.com/"
+        result.posts[1].tags == null
+        result.posts[1].contentType == "blog"
+    }
+
+    def "should throw exception for undefined error result"() {
+        given:
+        def filepath = packagePath.resolve("undefined_error_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        def ex = thrown(TwinglyException)
+        ex.message == "resultType:failure, message:Something went wrong."
+    }
+
+    def "should throw exception for unauthorized api key result"() {
+        given:
+        def filepath = packagePath.resolve("unauthorized_api_key_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        def ex = thrown(TwinglyException)
+        ex.message == "resultType:failure, message:The API key does not grant access to the Search API."
+    }
+
+    def "should throw exception for service unavailable result"() {
+        given:
+        def filepath = packagePath.resolve("service_unavailable_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        def ex = thrown(TwinglyException)
+        ex.message == "resultType:failure, message:Authentication service unavailable."
+    }
+
+
+    def "should throw exception for non existent api key result"() {
+        given:
+        def filepath = packagePath.resolve("nonexistent_api_key_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        def ex = thrown(TwinglyException)
+        ex.message == "resultType:failure, message:The API key does not exist."
+    }
+
+    def "should throw exception for non xml result"() {
+        given:
+        def filepath = packagePath.resolve("non_xml_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        def ex = thrown(TwinglyException)
+        ex.message == "Unable to process request"
+        assert ex.cause instanceof UnmarshalException
+    }
+
+    def "should parse minimal valid result"() {
+        given:
+        def filepath = packagePath.resolve("minimal_valid_result.xml")
+        def is = this.getClass().getClassLoader().getResourceAsStream(filepath.toString())
+        when:
+        def result = is.withReader("UTF-8", {
+            r -> return client.unmarshalXmlForResult(r)
+        })
+        then:
+        result.numberOfMatchesReturned == 1
+        result.numberOfMatchesTotal == 1
+        result.secondsElapsed == 0.148d
+        result.posts.size() == 3
+        result.posts[0].url == "http://oppogner.blogg.no/1409602010_bare_m_ha.html"
+        result.posts[0].blogName == "oppogner"
+        result.posts[0].blogUrl == "http://oppogner.blogg.no/"
+        result.posts[0].tags.size() == 1
+
+        result.posts[1].url == "http://www.skvallernytt.se/hardtraning-da-galler-swedish-house-mafia"
+        result.posts[1].blogName == "Skvallernytt.se"
+        result.posts[1].blogUrl == "http://www.skvallernytt.se/"
+        result.posts[1].tags.size() == 5
+
+        result.posts[2].url == "http://didriksinspesielleverden.blogg.no/1359472349_justin_bieber.html"
+        result.posts[2].blogName == "Didriksinspesielleverden"
+        result.posts[2].blogUrl == "http://didriksinspesielleverden.blogg.no/"
+        result.posts[2].tags == null
+    }
+
     def "client should change user agent"() {
         given:
         def newUserAgent = "newUserAgent"
-        def client = new UrlConnectionClient()
         when:
         client.setUserAgent(newUserAgent)
         def result = client.getUserAgent()
@@ -44,7 +185,6 @@ class UrlConnectionClientSpockTest extends Specification {
 
     def "client should throw IO exception that should be wrapped into Twingly Exception"() {
         given:
-        def client = Spy(UrlConnectionClient)
         def query = "http://twingly.com/"
         when:
         def result = client.makeRequest(query)
@@ -58,7 +198,6 @@ class UrlConnectionClientSpockTest extends Specification {
 
     def "client should throw JAXB exception that should be wrapped into Twingly Exception"() {
         given:
-        def client = Spy(UrlConnectionClient)
         def query = "http://twingly.com/"
         when:
         def result = client.makeRequest(query)
@@ -72,7 +211,6 @@ class UrlConnectionClientSpockTest extends Specification {
 
     def "should process query, return blog stream and throw exception"() {
         given:
-        def client = Spy(UrlConnectionClient)
         def query = "http://twingly.com/"
         def blogStream = new BlogStream()
         def operationResult = new OperationResult()
@@ -88,7 +226,6 @@ class UrlConnectionClientSpockTest extends Specification {
 
     def "should process query and return result isntance"() {
         given:
-        def client = Spy(UrlConnectionClient)
         def query = "http://twingly.com/"
         when:
         def result = client.makeRequest(query)
@@ -100,10 +237,7 @@ class UrlConnectionClientSpockTest extends Specification {
         assert result instanceof Result
     }
 
-
     def "client should create JAXB context"() {
-        given:
-        def client = new UrlConnectionClient()
         when:
         def context = client.getJAXBContext()
         then:
@@ -113,7 +247,6 @@ class UrlConnectionClientSpockTest extends Specification {
     def "should throw TwinglyException with MalformedURL"() {
         given:
         def malformedUrl = "qqq"
-        def client = new UrlConnectionClient()
         when:
         client.getUrl(malformedUrl)
         then:
@@ -125,12 +258,10 @@ class UrlConnectionClientSpockTest extends Specification {
     def "should create URL"() {
         given:
         def url = "http://google.com/"
-        def client = new UrlConnectionClient()
         when:
         def result = client.getUrl(url)
         then:
         result != null
     }
-
 
 }
